% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasq.R
\docType{data}
\name{datasq}
\alias{datasq}
\title{Magnetometer Observations of Station ABG, PHU, TUC and FRD from March to April 2001}
\format{A data frame containing 87840 observations for each four stations.}
\source{
\href{http://www.intermagnet.org/data-donnee/download-eng.php}{INTERMAGNET}
}
\usage{
data(datasq)
}
\description{
The H-component of the Ground-based magnetometer records from four stations, Alibag (ABG), Phuthuy (PHU), Tucson (TUC) and  Fredericksburg (FRD), used to estimate Sq variation. 
This data frame has 87840 rows, corresponding to a two months interval (March-April 2001), and 4 columns corresponding 
to each station.
}
\examples{
data(datasq)
head(datasq)
# plot of observations in station ABG
plot(datasq[,1])
}
\keyword{datasets}

