% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magnetic.latitude.R
\name{magnetic.latitude}
\alias{magnetic.latitude}
\title{Magnetic latitude}
\usage{
magnetic.latitude(col = 68.68, lam = 202)
}
\arguments{
\item{col}{Colatitude of a station}

\item{lam}{Longitude of a station}
}
\description{
Returns cos of magnetic latitude used in Dst. This function is used in SAIndex to adjust for the magnetic latitude.
}
\seealso{
\code{\link{SAIndex}}
}

