% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\docType{data}
\name{record}
\alias{record}
\title{Magnetometer Observations of Station HER, KAK, HON and SJG from March to April 2001}
\format{A data frame containing 87840 observations of H-component for four stations.}
\source{
\href{http://www.intermagnet.org/data-donnee/download-eng.php}{INTERMAGNET}
}
\usage{
data(record)
}
\description{
The H-component of the Ground-based magnetometer records four stations,Hermanus (HER), Kakioka (KAK), Honolulu (HON) and San Juan (SJG), used to 
estimate WISA and preindex. This data frame has 87840 rows, corresponding to a 
two months (March to April 2001) of 1-minute data, and 4 columns corresponding to each station.
}
\examples{
data(record)
head(record)

#plot of observations from station HER
plot(record[,1])
}
\keyword{datasets}

